# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Proker API",
        "description": "Dokumentasi API untuk Sistem Manajemen Program Kerja. Mencakup Authentication, Menu Management, Role & Group Management, dan Permission Management.",
        "contact": {
            "email": "admin@manajemenproker.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sys-groups": {
            "get": {
                "tags": [
                    "Sys Groups"
                ],
                "summary": "Daftar Grup / Role (Paginated)",
                "description": "Mengambil daftar seluruh grup pengguna dengan pagination.",
                "operationId": "6c465652a46f0fc6c09416940263c0fc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Halaman aktif",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_sys_group": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Super Admin"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Administrator utama"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "page": {
                                            "type": "integer"
                                        },
                                        "page_size": {
                                            "type": "integer"
                                        },
                                        "total_page": {
                                            "type": "integer"
                                        },
                                        "total_records": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sys Groups"
                ],
                "summary": "Tambah Grup Baru",
                "description": "Membuat grup / role pengguna baru.",
                "operationId": "bd7538184ba150efc58853fa1968906f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Editor"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Role untuk editor konten"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grup berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sys-groups/{id}": {
            "get": {
                "tags": [
                    "Sys Groups"
                ],
                "summary": "Detail Grup",
                "description": "Menampilkan detail satu grup berdasarkan ID.",
                "operationId": "f0cd24b2052c404d00fc5a76ead810d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Grup",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sys Groups"
                ],
                "summary": "Update Grup",
                "description": "Memperbarui data grup yang sudah ada.",
                "operationId": "17e1306f6d5112ce23ca6161d4281b15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Grup",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Editor Updated"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi baru"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sys Groups"
                ],
                "summary": "Hapus Grup",
                "description": "Menghapus grup berdasarkan ID.",
                "operationId": "5102427df5fabd8885af15b7c5c6d243",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Grup",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "data deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer {access_token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}